+++ app/utils/common/service-item.js
@@ -297,6 +297,11 @@ const CUSTOM_VERSIONS_FIELDSET = {
                 label: 'service_version.fields.in_catalogue'
               }
             ),
+            field(
+              'visible_to_marketplace', {
+                label: 'service_version.fields.visible_to_marketplace'
+              }
+            ),
           ]
         }
       }